+++ CHANGELOG.md
@@ -60,6 +60,7 @@ https://github.com/nwnxee/unified/compare/build8193.20...build8193.21
 - Creature: {Get/Set}Height()
 - Creature: {Get/Set}HitDistance()
 - Creature: {Get/Set}PreferredAttackDistance()
+- Creature: {Get/Set}BypassEffectImmunity()
 - Encounter: GetGeometry()
 - Encounter: SetGeometry()
 - Encounter: {Get/Set}CanReset()